Small fixes required to make it work on x86_64 architecture.
1. The comparision of ULONG field against -1UL is invalid on x86_64,
since the -1UL equals 0xffffffffffffffff and does not fit into ULONG.

2. The oop.library used utility.library but never opened it. Changed.


git-svn-id: https://svn.aros.org/svn/aros/trunk@26484 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ rom/oop/getattrbase.c
@@ -76,7 +76,7 @@
     	/* If so, it has been stored in the hashtable, and we have 
     	** to return the same numeric ID now.
 	*/
-	if (idb->attrbase == -1UL)
+	if (idb->attrbase == (ULONG)-1)
 	{
 	    /* The AttrBase has not yet been inited with ObtainAttrBase.
 	       I COULD init the attrbase now with the line below,
